requirements.txt
appdirs==1.4.4
cycler==0.11.0
dpkt==1.9.8
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.3
numpy==1.21.6
packaging==21.3
pandas==1.1.5
Pillow==9.3.0
psycopg2==2.9.5
py==1.11.0
pyparsing==3.0.9
pyshark==0.5.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
scapy==2.4.5
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
six==1.16.0
threadpoolctl==3.1.0
typing_extensions==4.4.0
